feat(nodeinfo): update NodeInfo struct to include detailed resource metrics and remove clippy errors
